fixed ailments page. Got data to display properly. Styled it as well as I could
