--- application/src/main/resources/static/js/react-client/src/pages/ailment-screen/style.module.scss
@@ -1,54 +1,0 @@
-.dropDwnBtn {
-    background-color: #045aaa;
-    color: white;
-    padding: 16px;
-    font-size: 16px;
-    border: none;
-  }
-  
-.dropdown {
-    position: relative;
-    display: inline-block;
-  }
-  
-  
-.ailmentText {
-    display: none;
-    position: absolute;
-    background-color: #f1f1f1;
-    min-width: 160px;
-    box-shadow: 0px 8px 16px 0px rgba(0,0,0,0.2);
-    z-index: 1;
-  }
-.treatmentText {
-    display: none;
-    position: absolute;
-    background-color: #f1f1f1;
-    min-width: 160px;
-    box-shadow: 0px 8px 16px 0px rgba(0,0,0,0.2);
-    z-index: 1;
-  }
-  
-
-.ailmentText a {
-    color: black;
-    padding: 12px 16px;
-    text-decoration: none;
-    display: block;
-  }
-.treatmentText a {
-    color: black;
-    padding: 12px 16px;
-    text-decoration: none;
-    display: block;
-  }
-  
-
-.ailmentText a:hover {background-color: #ddd;}
-.treatmentText a:hover {background-color: #ddd;}
-  
-.dropdown:hover .ailmentText {display: block;}
-.dropdown:hover .treatmentText {display: block;}
-  
-.dropdown:hover .dropDwnBtn {background-color: #3e858e;}
-
